requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
python-docx>=1.1.0
pydantic>=2.0.0
qdrant-client>=1.12.0
# Google Gemini
google-genai>=0.2.0
# OpenAI
openai>=1.50.0
# Test dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0
reportlab>=4.0.0